renamed module to less generic name
