+++ wisecondorX/triarray.py
